Adds thermometer, refactoring accessories.
Via adding a thermometer sensor, this addresses the problem with the
existing implementation and extending the available accessories.
By splitting accessories into different classes, SRP is better followed
and makes this more extendable.

Looking up which Accessory belongs to which model is done in the
Platform, and should close #18.

Thermometer is added as an accessory, closes #17.

Additions to test cases to mock listSensors as well are necessary to
accomodate for this, as well as bumping the Travis used node version.